requirements.txt
flask
google-generativeai
python-dotenv
pillow
groq